+++ packages/partials/concurrent/src/index.ts
@@ -18,3 +18,4 @@ export * from './Classes/ThrottleController';
 export * from './Classes/DebounceController';
 export * from './Classes/SlideWindowCounter';
 export * from './Classes/FiberPool';
+export * from './Classes/ManualBreaker';