+++ CHANGES.md
@@ -6,6 +6,7 @@
 - feat(pkg:test): added utility method `asyncThrows`
 - feat(pkg:ts-types): added helper type `IJsonSafeValue`
 - feat(pkg:ts-types): added helper type `IMaybeArray`
+- feat(pkg:concurrent): added utility class `ManualBreaker`
 
 ## v1.3.0